feat: add app: org.torproject.android
This change adds Orbot (stable and beta) to the application list. It is
not readily obvious what track the maintainers of the repository
consider to be "stable" -- going back several years, the only release
tags that are used include either "BETA" or "RC".

Even though "RC" (Release Candidate) releases are not often equated to
"stable" releases in the wider software world, these configurations make
the assumption that for the purpose of this application, "RC" indicates
"stable" from the maintainers.

Closes: imranr98/apps.obtainium.imranr.dev#498