[PATCH] Add button to go to the page showing the apk sha256 signature
Adds a `Show apk signature page`

When downloading apps through https://apps.obtainium.imranr.dev, finding the signing key can be not straight forward.
This adds a possible button to go to a page with the signing key available.
Some links were added to show how it would appear and start adding this informations to configurations